+++ CHANGES.md
@@ -1,3 +1,10 @@
+
+# 6.1.0 (TBD)
+
+* add `width`, `height` and `count` properties in `MosaicMethodBase`
+* make sure we mosaic ImageData/PointData with same number of bands
+* resize `ImageData.array` to the first asset's width/height in `mosaic_reader`
+
 # 6.0.2 (2023-08-21)
 
 * Update `data_as_image` to return masked values (author @JackDunnNZ, https://github.com/cogeotiff/rio-tiler/pull/635)